refactor: reorganize test structure and fix empty test files
- Reorganize all test files into logical directory structure:
  - core/ for foundational UI components
  - data-display/ for display components
  - forms/ for form components
  - layout/ for layout components
  - pages/ for page-level integration tests
- Add placeholder test suites for empty test files to prevent Vitest errors
- Move SocketProvider to components/providers/ directory
- Update imports in layout.tsx accordingly
- Fix 'No test suite found in file' errors by converting TODO comments to proper it.todo() tests
