+++ src/__tests__/components/ui/core/AnimationHeading.test.tsx
@@ -1,0 +1,9 @@
+import { describe, it } from 'vitest';
+
+describe('AnimationHeading', () => {
+  it.todo('should render heading with animation');
+  it.todo('should support different heading levels');
+  it.todo('should handle animation props correctly');
+  it.todo('should forward ref correctly');
+  it.todo('should support custom className');
+});
