+++ src/__tests__/components/ui/core/PageContainer.test.tsx
@@ -1,0 +1,9 @@
+import { describe, it } from 'vitest';
+
+describe('PageContainer', () => {
+  it.todo('should render container with proper layout');
+  it.todo('should support different container variants');
+  it.todo('should handle responsive behavior');
+  it.todo('should support custom className');
+  it.todo('should forward ref correctly');
+});
