+++ src/__tests__/components/ui/forms/Checkbox.test.tsx
@@ -1,0 +1,9 @@
+import { describe, it } from 'vitest';
+
+describe('Checkbox', () => {
+  it.todo('should render checkbox component');
+  it.todo('should handle checked state changes');
+  it.todo('should support disabled state');
+  it.todo('should support custom className');
+  it.todo('should forward ref correctly');
+});
