+++ src/__tests__/pages/home/HomePage.integration.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from '@testing-library/react';
 import { userEvent } from '@testing-library/user-event';
 import { describe, it, expect } from 'vitest';
-import { renderWithProviders } from '../setupTests';
+import { renderWithProviders } from '../../setupTests';
 import HomePage from '@/app/page';
 
 describe('HomePage Integration Tests', () => {
