+++ src/__tests__/utils/hooks/useAnimation.test.tsx
@@ -1,0 +1,9 @@
+import { describe, it } from 'vitest';
+
+describe('useAnimation', () => {
+  it.todo('should initialize with default animation state');
+  it.todo('should handle animation triggers correctly');
+  it.todo('should manage animation lifecycle');
+  it.todo('should support different animation types');
+  it.todo('should cleanup on unmount');
+});
